ItemSlot stuff

+++ incremental_affix/src/main.rs
@@ -3,6 +3,8 @@ mod game;
 
 use engine::drop_table;
 use engine::item;
+use game::modifiers::GameModifierKind;
+use game::player::Player;
 
 fn main() {
     let implicits = game::item::initialize_implicits();
@@ -15,26 +17,34 @@ fn main() {
     let bases_table = drop_tables.get::<drop_table::StorageBasesDropTable>().unwrap();
     let base_ix = bases_table.random();
 
-    let mut storage = item::AffixiveItem::new(&bases, &implicits, base_ix, item::Quality::Quality(1));
+    let mut random_item = item::AffixiveItem::new(&bases, &implicits, base_ix, item::Quality::Quality(1));
     
-    let suffix_or_prefix_table = drop_tables.get::<drop_table::SuffixOrPrefixDropTable>().unwrap();
-    match suffix_or_prefix_table.random() {
-        drop_table::SuffixOrPrefix::Prefix => {
-            let prefix_table = drop_tables.get::<drop_table::InventoryModifierPrefixes>().unwrap();
-            let mut prefix = prefix_table.random();
-            prefix.randomize_actual();
+    if random_item.tags.contains(&item::AffixiveItemTag::Inventory) {
+        let suffix_or_prefix_table = drop_tables.get::<drop_table::SuffixOrPrefixDropTable>().unwrap();
+        match suffix_or_prefix_table.random() {
+            drop_table::SuffixOrPrefix::Prefix => {
+                let prefix_table = drop_tables.get::<drop_table::InventoryModifierPrefixes<GameModifierKind>>().unwrap();
+                let mut prefix = prefix_table.random();
+                prefix.randomize_actual();
 
-            let _ = storage.try_push_prefix(prefix);
-        },
+                let _ = random_item.try_push_prefix(prefix);
+            },
 
-        drop_table::SuffixOrPrefix::Suffix => {
-            let suffix_table = drop_tables.get::<drop_table::InventoryModifierSuffixes>().unwrap();
-            let mut suffix = suffix_table.random();
-            suffix.randomize_actual();
+            drop_table::SuffixOrPrefix::Suffix => {
+                let suffix_table = drop_tables.get::<drop_table::InventoryModifierSuffixes<GameModifierKind>>().unwrap();
+                let mut suffix = suffix_table.random();
+                suffix.randomize_actual();
 
-            let _ = storage.try_push_suffix(suffix);
+                let _ = random_item.try_push_suffix(suffix);
+            }
         }
     }
 
-    println!("{}", storage.display(&bases[..]));
+    println!("{}", random_item.display(&bases[..]));
+
+    let mut player = Player::default();
+    match player.try_equipping(random_item, &bases) {
+        Ok(_) => println!("Inventory Volume: {}", player.get_inventory_volume()),
+        Err(_) => { panic!("Player cannot equip?" )},
+    }
 }
